+++ packages/token-bridge-sdk/src/util/fetchETHWithdrawalsFromSubgraph.ts
@@ -1,6 +1,3 @@
-import { Provider, BlockTag } from '@ethersproject/providers'
-import { L2ToL1MessageReader } from '@arbitrum/sdk'
-
 import { getETHWithdrawals as graph_getETHWithdrawals } from './graph'
 
 /**
@@ -29,33 +26,5 @@ export function fetchETHWithdrawalsFromSubgraph({
     fromBlock,
     toBlock,
     String(l1NetworkId)
-  )
-}
-
-/**
- * Fetches initiated ETH withdrawals from event logs in range of [fromBlock, toBlock].
- *
- * @param query Query params
- * @param query.address Account address
- * @param query.fromBlock Start at this block number (including)
- * @param query.toBlock Stop at this block number (including)
- * @param query.l2Provider Provider for the L2 network
- */
-export function fetchETHWithdrawalsFromEventLogs({
-  address,
-  fromBlock,
-  toBlock,
-  l2Provider
-}: {
-  address: string
-  fromBlock: BlockTag
-  toBlock: BlockTag
-  l2Provider: Provider
-}) {
-  return L2ToL1MessageReader.getEventLogs(
-    l2Provider,
-    { fromBlock, toBlock },
-    undefined,
-    address
   )
 }
